Add passwd (build-)dependency for chpasswd and newusers usage
The build dependency is only to let configure.ac pick up the right path
for the utilities (which is in their default path now, but who known in future).
The runtime dependency is added to cockpit-bridge which is what will execute them.

Also note that configure.ac checks for a newusers feature which we don't
have in Debian.