Move to copyright template, split to separate License: paragraphs
ftpmasters did not accept COPYING.bower, so stop installing it, and
replace the generic paragraph for it in debian/copyright with a
placeholder. Rename it to debian/copyright.template.

Add debian/build-copyright which builds debian/copyright from
copyright.template and a machine audit of bower_components/*. This
currently includes a few quirks due to missing licenses and copyrights,
the corresponding upstream bug reports are mentioned in comments.